thiet ke trang chon ghe va them 1 so thong tin khac

+++ ViewCinema/CinemaProject/CinemaProject/Models/Ghe.cs
@@ -2,8 +2,10 @@
 {
     public class Ghe
     {
-        public string IdGhe { get; set; }     
-        public string LoaiGhe { get; set; }    
-        public decimal Gia { get; set; }     
+        public string IdGhe { get; set; }
+
+        public string LoaiGhe { get; set; }
+
+        public int Gia { get; set; }
     }
 }
